chore: refactor algolia indexer
