+++ apps/demo/src/pages/api/reindex.ts
@@ -1,8 +1,10 @@
-import { index } from '../../lib/indexer';
+import { indexer } from '@watheia/content-api';
+
+import { defaultSchema } from '../../lib/content-api';
 
 export default async function handler(req, res) {
   try {
-    const result = await index();
+    const result = await indexer(defaultSchema);
     res.status(200).send({ result });
   } catch (err) {
     console.error(err);
