+++ libs/content-api/package.json
@@ -4,13 +4,15 @@
   "dependencies": {
     "@iarna/toml": "2.2.5",
     "@swc/helpers": "~0.5.6",
+    "@watheia/content-helpers": "0.0.1",
     "@watheia/content-model": "0.0.1",
+    "algoliasearch": "^4.23.2",
     "fast-glob": "^3.3.2",
+    "front-matter": "^4.0.2",
+    "jose": "^3.20.4",
     "js-yaml": "4.1.0",
     "lodash": "4.17.21",
-    "@watheia/content-helpers": "0.0.1",
-    "front-matter": "^4.0.2",
-    "jose": "^3.20.4"
+    "marked": "^12.0.1"
   },
   "type": "commonjs",
   "main": "./src/index.js",
