+++ libs/content-api/src/index.ts
@@ -1,6 +1,5 @@
-export { LocalContentApi } from './content-api';
-export type {
-  ContentApi,
-  LocalContentSchema as LocalContentSchema,
-  MetaTag,
-} from './lib/content-api.types';
+export { LocalContentApi } from './lib/content-api';
+export { indexer } from './lib/indexer';
+export { PlainTextRenderer } from './lib/markdown-plaintext';
+
+export type { ContentApi, LocalContentSchema } from './lib/content-api.types';
