+++ libs/content-api/src/lib/content-api.ts
@@ -1,5 +1,5 @@
-import type { ContentCache, LocalContentSchema } from './lib/content-api.types';
-import { resolveContent, resolveStaticPaths, resolveStaticProps } from './lib/resolver';
+import type { ContentCache, LocalContentSchema } from './content-api.types';
+import { resolveContent, resolveStaticPaths, resolveStaticProps } from './resolver';
 
 export class LocalContentApi {
   public static async create(schema: LocalContentSchema): Promise<LocalContentApi> {
