+++ libs/content-helpers/src/index.ts
@@ -1,3 +1,4 @@
+export * from './lib/consts';
 export * from './lib/content-helpers';
 export * from './lib/data-utils';
 export * from './lib/get-data-attrs';
